Fix todo delete functionality
In case of a non existent id, an error would be thrown

+++ backend/src/main/java/com/darius/todoapp/dao/TodoDAOImpl.java
@@ -47,7 +47,8 @@ public class TodoDAOImpl implements TodoDAO {
 
         Todo theTodo = entityManager.find(Todo.class, theId);
 
-        entityManager.remove(theTodo);
+        if(theTodo != null)
+            entityManager.remove(theTodo);
     }
 
     @Override
